Create mlpnn_test0.py
First attempt at making the model work. Signals fed in as absolute value of Fourier transforms as I cannot put complex data in so I am losing quite a lot of information.  We are definitely able to feed in the raw signals but will have to make all snippets the same length. It takes a while to run. Current accuracy: ~70%.