Update trainingDataSelector.py
Now ignore the first 10 entries of Fourier transform array, it seems to be disproportionately large compared to everything else. No longer want to go up to index 1000 as it seems most of the time Fourier contributions are non-existent and also this allows us to get use some of the shorter snippets which didn't have enough entries in the array to get 1000 elements in FT array.