Calculating and saving Precision and Recall statistics
Having done some reading on the different metrics we can use to see how our model is performing (https://scikit-learn.org/stable/modules/model_evaluation.html), it looks like precision and recall could be quite useful as it allows us to see how model is performing for a certain condition. Added functionality to calculate this for each iteration and to do some statistics on our results. Again, everyone run this and we will combine the results on Monday.